[#8] [Backend] Fix file name, remove unused imports, use final keyword, rename parameter
