+++ lib/usecases/login_use_case.dart
@@ -1,9 +1,9 @@
-import 'package:flutter_survey/model/AuthToken.dart';
 import 'package:injectable/injectable.dart';
 
 import '../api/exception/network_exceptions.dart';
 import '../api/repository/authentication_repository.dart';
 import '../local/local_storage.dart';
+import '../model/auth_token.dart';
 import 'base/base_use_case.dart';
 
 class LoginInput {
@@ -21,9 +21,9 @@ class LoginUseCase extends UseCase<void, LoginInput> {
   const LoginUseCase(this._repository, this._localStorage);
 
   @override
-  Future<Result<void>> call(LoginInput input) {
+  Future<Result<void>> call(LoginInput params) {
     return _repository
-        .logIn(email: input.email, password: input.password)
+        .logIn(email: params.email, password: params.password)
         .then((token) => saveTokens(token))
         .onError<NetworkExceptions>(
             (err, stackTrace) => Failed(UseCaseException(err)));
